pyblish_ci/requirements.txt
flask
flask-restful
requests
pyyaml